Fix Trips GET and (maybe) POST Endpoints
Massive overhaul and rework of all classes involved with booking a trip. Will make life easier for future commits.

+++ src/main/java/famu/edu/a1travel/Model/Trains.java
@@ -1,11 +1,8 @@
 package famu.edu.a1travel.Model;
 import com.google.cloud.Timestamp;
-import com.google.protobuf.util.Timestamps;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.text.ParseException;
 
 @Data
 @NoArgsConstructor
@@ -17,7 +14,7 @@ public class Trains {
     private String departLocation;
     private String duration;
     private String status;
-    private float price;
+    private Double price;
     private int stops;
 
     /*public void setArriveTime(String arriveTime) throws ParseException {
